Replace 'Float' with 'Double' in ScrollConverter API
On very large lists (where the total content size exceeds > 1_000_000
px), converter loses precision due to the use of float. At this
magnitude, a 32-bit float no longer has enough significant digits to
represent fractional pixel values, because almost all precision is
consumed by the integer part. As a result, small rounding errors
accumulate at the bottom of the list. This is a limitation of
floating-point representation and it can't be avoided as long as we
operate on values larger than ~1e6 using float.

This CL is a preparation step for the upcoming fix in the next CL.
Without this change, I cannot write tests that can properly check big
numbers.

Test: 484296898

+++ xr/glimmer/glimmer/src/androidTest/kotlin/androidx/xr/glimmer/list/ListAnimateScrollToItemTest.kt
@@ -208,7 +208,7 @@ class ListAnimateScrollToItemTest(private val testCase: ScrollMovementTestCase) 
 
     private val ListState.focusLinePosition: Float
         get() {
-            val focusScroll = requireNotNull(autoFocusState.properties).focusScroll
+            val focusScroll = requireNotNull(autoFocusState.properties).focusScroll.toFloat()
             val startPadding = layoutInfo.beforeContentPadding
             return startPadding + focusScroll
         }
